fix: consolidate submit response mapping for pkg-file-count CI
Mergeability: fold submitWorkResponseFromResult into submit_work_items.go
and move its test into server_submit_work_test.go so pkg/api stays within
the 15-file package limit after story 001 additions.
