--- pkg/api/submit_response.go
@@ -1,27 +1,0 @@
-package api
-
-import (
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-)
-
-func submitWorkResponseFromResult(result interfaces.WorkRequestSubmitResult, sessionID string) factoryapi.SubmitWorkResponse {
-	resp := factoryapi.SubmitWorkResponse{
-		TraceId:   result.TraceID,
-		RequestId: result.RequestID,
-		Accepted:  result.Accepted,
-	}
-	if result.WorkID != "" {
-		resp.WorkId = &result.WorkID
-	}
-	if result.Name != "" {
-		resp.Name = &result.Name
-	}
-	if result.WorkTypeName != "" {
-		resp.WorkTypeName = &result.WorkTypeName
-	}
-	if sessionID != "" {
-		resp.SessionId = &sessionID
-	}
-	return resp
-}
