+++ pkg/api/submit_work_items.go
@@ -237,3 +237,24 @@ func requiredSubmitWorkItemType(fields map[string]json.RawMessage, prefix string
 		return "", requestFieldValidationError{message: fmt.Sprintf("%stype must be one of text, image, video, audio, or document", prefix)}
 	}
 }
+
+func submitWorkResponseFromResult(result interfaces.WorkRequestSubmitResult, sessionID string) factoryapi.SubmitWorkResponse {
+	resp := factoryapi.SubmitWorkResponse{
+		TraceId:   result.TraceID,
+		RequestId: result.RequestID,
+		Accepted:  result.Accepted,
+	}
+	if result.WorkID != "" {
+		resp.WorkId = &result.WorkID
+	}
+	if result.Name != "" {
+		resp.Name = &result.Name
+	}
+	if result.WorkTypeName != "" {
+		resp.WorkTypeName = &result.WorkTypeName
+	}
+	if sessionID != "" {
+		resp.SessionId = &sessionID
+	}
+	return resp
+}
